+++ crates/ruma-client-api/src/session/sso_login_with_provider.rs
@@ -83,17 +83,21 @@ pub mod v3 {
 
     #[cfg(all(test, feature = "client"))]
     mod tests {
-        use ruma_common::api::{MatrixVersion, OutgoingRequest as _, SendAccessToken};
+        use ruma_common::api::{
+            MatrixVersion, OutgoingRequest as _, SendAccessToken, SupportedVersions,
+        };
 
         use super::Request;
 
         #[test]
         fn serialize_sso_login_with_provider_request_uri() {
+            let supported =
+                SupportedVersions { versions: [MatrixVersion::V1_1].into(), features: Vec::new() };
             let req = Request::new("provider".to_owned(), "https://example.com/sso".to_owned())
                 .try_into_http_request::<Vec<u8>>(
                     "https://homeserver.tld",
                     SendAccessToken::None,
-                    &[MatrixVersion::V1_1],
+                    &supported,
                 )
                 .unwrap();